refactor: split out all pages into their own files
