+++ src/pages/mod.rs
@@ -1,0 +1,7 @@
+pub mod auth;
+pub mod index;
+pub mod login;
+pub mod logo;
+pub mod qr;
+pub mod register_attendance;
+pub mod session_week;
