+++ src/pages/qr.rs
